Module type Internal_event.EVENT_DEFINITION

Parameters defining an inspectable type of events.

type t

val section : Section.t option

Defines an optional section for the event.

Warning None is only for legacy events and should not be used in new code.

val name : string

Defines the identifier for the event. Names should be unique and are restricted to alphanumeric characters or ".@-_+=,~".

val doc : string

A display-friendly text which describes what the event means.

val pp : short:bool -> Stdlib.Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val level : t -> level

Return the preferred level for a given event instance.