Module Tezos_event_logging__Internal_event.Debug_event

The debug-event is meant for emitting (temporarily) semi-structured data in the event stream.

type t = {

message : string;

attachment : Data_encoding.Json.t;

}

val make : ?⁠attach:Data_encoding.Json.t -> string -> unit -> t

include EVENT with type t := t

include EVENT_DEFINITION

type t

val section : Section.t option

Defines an optional section for the event.

Warning None is only for legacy events and should not be used in new code.

val name : string

Defines the identifier for the event. Names should be unique and are restricted to alphanumeric characters or ".@-_+=,~".

val doc : string

A display-friendly text which describes what the event means.

val pp : short:bool -> Stdlib.Format.formatter -> t -> unit

val encoding : t Data_encoding.t

val level : t -> level

Return the preferred level for a given event instance.

val emit : ?⁠section:Section.t -> (unit -> t) -> unit Tezos_error_monad.Error_monad.tzresult Lwt.t

Output an event of type t, if no sinks are listening the function won't be applied.