Module Tezos_event_logging__Internal_event.Lwt_log_sink

The lwt-sink outputs pretty-printed renderings of events to the lwt-log logging framework (see the Lwt_log_core module).

It is activated by default in Internal_event_unix.Configuration.default (in any case it can be activated with TEZOS_EVENTS_CONFIG="lwt-log://". To configure further how the sink outputs to a file or the user's terminal, one needs to use the TEZOS_LOG variable (see also the module Lwt_log_sink_unix).

val uri_scheme : string